feat(shared): notebook cleanup
